requirements.txt
tabula-py
pandas
PyPDF2
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scipy
